+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 0.3.0 / 2018-04-12
 ==================
 - Add GSD ID to output data
+- Map telephone number to `contacts.telephoneNumber`
+- Map service type to `serviceType`
+- Map opening times to `openingTimes.description`
 
 0.2.0 / 2018-03-28
 ==================